# Improve speed and memory consumption of binned `PrecisionRecallCurve`