## What does this PR do?

Fixes #1492 by reverting the implementation of the update step of binary and multiclass PrecisionRecallCurve with a fixed number of thresholds to the implementation from torchmetrics 0.9.

I did not touch the multilabel case, because it is difficult to adapt the `ignore_idx` filtering logic, a first attempt of implementing it on a per-threshold basis was much slower, and multilabel classification is not common in semantic segmentation scenarios, which is where we typically have to deal with huge numbers of samples.

<details>
  <summary>Before submitting</summary>

- [x] Was this **discussed/agreed** via a Github issue? (no need for typos and docs improvements)
- [x] Did you read the [contributor guideline](https://github.com/Lightning-AI/metrics/blob/master/.github/CONTRIBUTING.md), Pull Request section?
- [ ] Did you make sure to **update the docs**?
- [ ] Did you write any new **necessary tests**?

</details>

<details>
  <summary>PR review</summary>

Anyone in the community is free to review the PR once the tests have passed.
If we didn't discuss your PR in Github issues there's a high chance it will not be merged.

</details>

## Performance Comparison

Hardware: NVIDIA RTX 3090

### `BinaryPrecisionRecallCurve.update()`

`thresholds=200`

|  #Samples  | Current Speed | New Speed | Current Memory | New Memory |
|-----------:|--------------:|----------:|---------------:|-----------:|
|        100 |        0.2 ms |   23.8 ms |         < 1 MB |     < 1 MB |
|      1,000 |        1.1 ms |   24.4 ms |          20 MB |     < 1 MB |
|     25,000 |       24.0 ms |   25.8 ms |         200 MB |     < 1 MB |
|    100,000 |      115.3 ms |   25.5 ms |         722 MB |       2 MB |
|  1,000,000 |    1,012.1 ms |   26.7 ms |       7,156 MB |       2 MB |
|  3,000,000 |    7,948.3 ms |   56.5 ms |      18,886 MB |      44 MB |

### `MulticlassPrecisionRecallCurve.update()`

`num_classes=100, thresholds=200`

|  #Samples  | Current Speed | New Speed | Current Memory | New Memory |
|-----------:|--------------:|----------:|---------------:|-----------:|
|         10 |        0.2 ms |   15.5 ms |          20 MB |     < 1 MB |
|        100 |        0.5 ms |   15.2 ms |          52 MB |     < 1 MB |
|      1,000 |        3.5 ms |   15.2 ms |         484 MB |       4 MB |
|     10,000 |       30.6 ms |   30.8 ms |       4,770 MB |      22 MB |
|     25,000 |       76.6 ms |   70.6 ms |      12,404 MB |      60 MB |

The implementation proposed here is not faster in all cases but always more memory efficient. In particular, it is up to 100x slower for few numbers of samples such as in a typical image-level classification scenario. However, it is up to 100x faster for scenarios with many samples per batch such as in semantic segmentation use-cases. In these scenarios, the extreme memory consumption of the previous implementation is the largest deal breaker.

## Did you have fun?

Make sure you had fun coding 🙃