@Callidior 2023-02-09
A more detailed speed benchmark on the binary case (time in ms):

![image](https://user-images.githubusercontent.com/7915048/217856074-2c7efb1f-4413-40f0-b5c2-0adb576e3da4.png)

- The loop implementation scales linearly with the number of thresholds but is almost insensitive to the number of samples.
- The vectorized implementation scales linearly with the number of samples.
- Keep in mind that memory consumption also scales almost linearly with number of samples for the vectorized implementation. The case in the bottom right corner takes 24 GB of VRAM, while the loop implementation has a constant memory consumption of 2 MB.

The documentation of `PrecisionRecallCurve` currently states;

> The implementation both supports calculating the metric in a non-binned but accurate version and a binned version that is less accurate but more memory efficient. Setting the thresholds argument to None will activate the non-binned version that uses memory of size O(n_samples) whereas setting the thresholds argument to either an integer, list or a 1d tensor will use a binned version that uses memory of size O(n_thresholds) (constant memory).

That is true for the metric's state but not for intermediate computations during the update step. The documentation does not mention speed but puts a promise on memory efficiency in the foreground, which is currently not kept.

With the current implementation, I don't see any benefit of using the binned version of the metric. If you have few samples, the accurate version (`thresholds=None`) is still very tractable and more accurate than the binned one. If you have many samples, both the binned and the accurate version of unusable due to high memory requirements and computation time.

@justusschock: Given that users can already switch between the accurate and the binned version, I don't think an explicit second switch for two different implementations of the binned version makes sense.