@SkafteNicki 2023-02-10
I would agree with @justusschock that the correct approach is implement some kind of heuristic that changes between `bincount` and a `for-loop` depending on how many samples and threshold the user are using. To settle on this heuristic, would it be possible for you @Callidior do share the benchmark script you have developed. In particular I am also interested in what the difference between doing this on CPU is (it seems you have only tried out GPU).